refactor: Update track mj event to helper function
